Compilation error fixed by moving iccsecurezero.c file to icc/
